requirements.txt
ipykernel
numpy
pytorch
pytorch-cuda
protobuf
pynvml
einops
ipywidgets
tqdm
scikit-learn
simple-parsing
plotly
datasets
sentencepiece
circuitsvis  # pip
transformer_lens  # pip
-e lib/elk  # pip
